move Void to xdr_codec
