+++ crates/nfs3_client/src/portmapper.rs
@@ -1,32 +1,10 @@
 use std::io::Cursor;
 
 use nfs3_types::portmap::{mapping, PMAP_PROG, PROGRAM, VERSION};
-use nfs3_types::xdr_codec::{Pack, PackedSize, Unpack};
+use nfs3_types::xdr_codec::{Pack, PackedSize, Unpack, Void};
 
 use crate::io::{AsyncRead, AsyncWrite};
 use crate::rpc::RpcClient;
-
-struct Void;
-
-impl<Out: std::io::Write> Pack<Out> for Void {
-    fn pack(&self, _buf: &mut Out) -> nfs3_types::xdr_codec::Result<usize> {
-        Ok(0)
-    }
-}
-
-impl PackedSize for Void {
-    const PACKED_SIZE: Option<usize> = Some(0);
-
-    fn count_packed_size(&self) -> usize {
-        0
-    }
-}
-
-impl<In: std::io::Read> Unpack<In> for Void {
-    fn unpack(_buf: &mut In) -> nfs3_types::xdr_codec::Result<(Void, usize)> {
-        Ok((Void, 0))
-    }
-}
 
 pub struct PortmapperClient<IO> {
     rpc: RpcClient<IO>,
